Refactored GloFAS class
