+++ analyses/bgd/trigger_development/utils/utils.py
@@ -11,7 +11,11 @@ path_mod = f"{Path(os.path.dirname(os.path.realpath(''))).parents[1]}/"
 sys.path.append(path_mod)
 
 from src.bgd import get_glofas_data as ggd
-from src.indicators.flooding.glofas import glofas
+from src.indicators.flooding.glofas.forecast import (
+    GlofasForecast,
+    GlofasReforecast,
+)
+from src.indicators.flooding.glofas.reanalysis import GlofasReanalysis
 
 pd.options.mode.chained_assignment = None
 
@@ -46,7 +50,7 @@ def get_glofas_df(
 
 
 def get_glofas_reanalysis(version: int = 3):
-    glofas_reanalysis = glofas.GlofasReanalysis()
+    glofas_reanalysis = GlofasReanalysis()
     da_glofas_reanalysis = glofas_reanalysis.read_processed_dataset(
         country_iso3=ggd.COUNTRY_ISO3, version=version
     )[STATION]
@@ -54,7 +58,7 @@ def get_glofas_reanalysis(version: int = 3):
 
 
 def get_glofas_forecast(version: int = 3, leadtimes: list = ggd.LEADTIMES):
-    glofas_forecast = glofas.GlofasForecast()
+    glofas_forecast = GlofasForecast()
     da_glofas_forecast_dict = {
         leadtime: glofas_forecast.read_processed_dataset(
             country_iso3=ggd.COUNTRY_ISO3,
@@ -70,7 +74,7 @@ def get_glofas_forecast(version: int = 3, leadtimes: list = ggd.LEADTIMES):
 def get_glofas_reforecast(
     version: int = 3, interp: bool = True, leadtimes: list = ggd.LEADTIMES
 ):
-    glofas_reforecast = glofas.GlofasReforecast()
+    glofas_reforecast = GlofasReforecast()
     da_glofas_reforecast_dict = {
         leadtime: glofas_reforecast.read_processed_dataset(
             country_iso3=ggd.COUNTRY_ISO3,
