+++ src/indicators/flooding/glofas/glofas.py
@@ -266,214 +266,21 @@ class Glofas:
         return xr.load_dataset(filepath)
 
 
-class GlofasReanalysis(Glofas):
-    def __init__(self, **kwargs):
-        super().__init__(
-            year_min=1979,
-            year_max=2020,
-            cds_name="cems-glofas-historical",
-            dataset=["consolidated_reanalysis"],
-            dataset_variable_name="dataset",
-            system_version_minor={2: 1, 3: 1},
-            date_variable_prefix="h",
-            **kwargs,
-        )
-
-    def download(
+class CoordsOutOfBounds(Exception):
+    def __init__(
         self,
-        country_iso3: str,
-        area: Area,
-        version: int = DEFAULT_VERSION,
-        year_min: int = None,
-        year_max: int = None,
+        station_name: str,
+        param_name: str,
+        coord_station: float,
+        coord_min: float,
+        coord_max: float,
     ):
-        year_min = self.year_min if year_min is None else year_min
-        year_max = self.year_max if year_max is None else year_max
-        logger.info(
-            f"Downloading GloFAS reanalysis v{version} for years {year_min} -"
-            f" {year_max}"
+        message = (
+            f"Station {station_name} has out-of-bounds {param_name} value of"
+            f" {coord_station} (GloFAS {param_name} ranges from {coord_min} to"
+            f" {coord_max})"
         )
-        for year in range(year_min, year_max + 1):
-            logger.info(f"...{year}")
-            super()._download(
-                country_iso3=country_iso3,
-                area=area,
-                year=year,
-                version=version,
-            )
-
-    def process(
-        self,
-        country_iso3: str,
-        stations: Dict[str, Station],
-        version: int = DEFAULT_VERSION,
-    ):
-        # Get list of files to open
-        logger.info(f"Processing GloFAS Reanalysis v{version}")
-        filepath_list = [
-            self._get_raw_filepath(
-                country_iso3=country_iso3,
-                version=version,
-                year=year,
-            )
-            for year in range(self.year_min, self.year_max + 1)
-        ]
-        # Read in the dataset
-        logger.info(f"Reading in {len(filepath_list)} files")
-
-        with xr.open_mfdataset(
-            filepath_list, engine="cfgrib", backend_kwargs={"indexpath": ""}
-        ) as ds:
-            # Create a new dataset with just the station pixels
-            logger.info("Looping through stations, this takes some time")
-            ds_new = _get_station_dataset(
-                stations=stations, ds=ds, coord_names=["time"]
-            )
-        # Write out the new dataset to a file
-        self._write_to_processed_file(
-            country_iso3=country_iso3,
-            version=version,
-            ds=ds_new,
-        )
-
-
-class GlofasForecastBase(Glofas):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def _download(
-        self,
-        is_reforecast: bool,
-        country_iso3: str,
-        area: Area,
-        leadtimes: List[int],
-        version: int = DEFAULT_VERSION,
-        split_by_month: bool = False,
-        split_by_leadtimes: bool = False,
-        year_min: int = None,
-        year_max: int = None,
-    ):
-        forecast_type = "reforecast" if is_reforecast else "forecast"
-        year_min = self.year_min[version] if year_min is None else year_min
-        year_max = self.year_max if year_max is None else year_max
-        logger.info(
-            f"Downloading GloFAS {forecast_type} v{version} for years"
-            f" {year_min} - {year_max} and lead time {leadtimes}"
-        )
-        for year in range(year_min, year_max + 1):
-            logger.info(f"...{year}")
-            month_range = range(1, 13) if split_by_month else [None]
-            for month in month_range:
-                leadtime_range = (
-                    leadtimes if split_by_leadtimes else [leadtimes]
-                )
-                for leadtime in leadtime_range:
-                    super()._download(
-                        country_iso3=country_iso3,
-                        area=area,
-                        version=version,
-                        year=year,
-                        month=month,
-                        leadtime=leadtime,
-                    )
-
-    def _process(
-        self,
-        is_reforecast: bool,
-        country_iso3: str,
-        stations: Dict[str, Station],
-        leadtimes: List[int],
-        version: int = DEFAULT_VERSION,
-        split_by_month: bool = False,
-        split_by_leadtimes: bool = False,
-        year_min: int = None,
-        year_max: int = None,
-    ):
-        forecast_type = "reforecast" if is_reforecast else "forecast"
-        year_min = self.year_min[version] if year_min is None else year_min
-        year_max = self.year_max if year_max is None else year_max
-        logger.info(
-            f"Processing GloFAS {forecast_type} v{version} for years"
-            f" {year_min} - {year_max} and lead time {leadtimes}"
-        )
-        month_range = range(1, 13) if split_by_month else [None]
-        leadtime_range = leadtimes if split_by_leadtimes else [leadtimes]
-        for leadtime in leadtime_range:
-            logger.info(f"For lead time {leadtime}")
-            # Get list of files to open
-            filepath_list = [
-                self._get_raw_filepath(
-                    country_iso3=country_iso3,
-                    version=version,
-                    year=year,
-                    month=month,
-                    leadtime=leadtime,
-                )
-                for year in range(year_min, year_max + 1)
-                for month in month_range
-            ]
-            # Read in both the control and ensemble perturbed forecast
-            # and combine
-            logger.info(f"Reading in {len(filepath_list)} files")
-            ds = self._read_in_ensemble_and_perturbed_datasets(
-                filepath_list=filepath_list
-            )
-            # Create a new dataset with just the station pixels
-            logger.info("Looping through stations, this takes some time")
-            coord_names = ["number", "time"]
-            if not split_by_leadtimes:
-                coord_names += ["step"]
-            ds_new = _get_station_dataset(
-                stations=stations,
-                ds=ds,
-                coord_names=coord_names,
-            )
-            # Write out the new dataset to a file
-            self._write_to_processed_file(
-                country_iso3=country_iso3,
-                version=version,
-                ds=ds_new,
-                leadtime=leadtime,
-            )
-
-
-class GlofasForecast(GlofasForecastBase):
-    def __init__(self, **kwargs):
-        super().__init__(
-            year_min={2: 2019, 3: 2020},
-            year_max=2020,
-            cds_name="cems-glofas-forecast",
-            dataset=["control_forecast", "ensemble_perturbed_forecasts"],
-            system_version_minor={2: 1, 3: 1},
-            dataset_variable_name="product_type",
-            **kwargs,
-        )
-
-    def download(self, *args, **kwargs):
-        super()._download(is_reforecast=False, *args, **kwargs)
-
-    def process(self, *args, **kwargs):
-        super()._process(is_reforecast=False, *args, **kwargs)
-
-
-class GlofasReforecast(GlofasForecastBase):
-    def __init__(self, **kwargs):
-        super().__init__(
-            year_min={2: 1999, 3: 1999},
-            year_max=2018,
-            cds_name="cems-glofas-reforecast",
-            dataset=["control_reforecast", "ensemble_perturbed_reforecasts"],
-            dataset_variable_name="product_type",
-            system_version_minor={2: 2, 3: 1},
-            date_variable_prefix="h",
-            **kwargs,
-        )
-
-    def download(self, *args, **kwargs):
-        super()._download(is_reforecast=True, *args, **kwargs)
-
-    def process(self, *args, **kwargs):
-        super()._process(is_reforecast=True, *args, **kwargs)
+        super().__init__(message)
 
 
 def expand_dims(
@@ -498,24 +305,7 @@ def expand_dims(
     return ds
 
 
-class CoordsOutOfBounds(Exception):
-    def __init__(
-        self,
-        station_name: str,
-        param_name: str,
-        coord_station: float,
-        coord_min: float,
-        coord_max: float,
-    ):
-        message = (
-            f"Station {station_name} has out-of-bounds {param_name} value of"
-            f" {coord_station} (GloFAS {param_name} ranges from {coord_min} to"
-            f" {coord_max})"
-        )
-        super().__init__(message)
-
-
-def _get_station_dataset(
+def get_station_dataset(
     stations: Dict[str, Station], ds: xr.Dataset, coord_names: List[str]
 ) -> xr.Dataset:
     # Check that lat and lon are in the bounds
