+++ src/indicators/flooding/glofas/utils.py
@@ -7,6 +7,11 @@ import xarray as xr
 from scipy.stats import rankdata
 
 from src.indicators.flooding.glofas import glofas
+from src.indicators.flooding.glofas.forecast import (
+    GlofasForecast,
+    GlofasReforecast,
+)
+from src.indicators.flooding.glofas.reanalysis import GlofasReanalysis
 from src.utils_general.statistics import (
     calc_crps,
     get_return_period_function_analytical,
@@ -19,7 +24,7 @@ logger = logging.getLogger(__name__)
 def get_glofas_reanalysis(
     country_iso3: str, version: int = glofas.DEFAULT_VERSION, **kwargs
 ) -> xr.Dataset:
-    glofas_reanalysis = glofas.GlofasReanalysis(**kwargs)
+    glofas_reanalysis = GlofasReanalysis(**kwargs)
     ds_glofas_reanalysis = glofas_reanalysis.read_processed_dataset(
         country_iso3=country_iso3, version=version
     )
@@ -37,9 +42,9 @@ def _get_glofas_forecast_base(
     **kwargs,
 ) -> xr.Dataset:
     if is_reforecast:
-        glofas_forecast = glofas.GlofasReforecast(**kwargs)
+        glofas_forecast = GlofasReforecast(**kwargs)
     else:
-        glofas_forecast = glofas.GlofasForecast(**kwargs)
+        glofas_forecast = GlofasForecast(**kwargs)
     if split_by_leadtimes:
         ds_glofas_forecast_dict = {
             leadtime: glofas_forecast.read_processed_dataset(
